VeloxRaster_get_data_type {velox}	R Documentation

Get data type of a VeloxRaster

Description

Get data type of a VeloxRaster

Details

Note that this method returns the data type of the raster, not the storage mode. Except in special cases, velox stores all raster data as double precision matrices.

Value

A character string denoting a GDAL data type.

[Package velox version 0.2.1 Index]